move search utils to general utils
